Move MapTab wiring and preserve sb.explore
